[LayoutNG] Avoid false positives for clearance.
Clearance inhibits margin collapsing and thus makes it possible to
resolve the BFC offset of its parent. In that regard we sometimes
incorrectly assumed that a child would get clearance. The only case
where we can tell for sure up front, is when there are adjoining floats
that match the value of the 'clear' property. In all other cases, we
first need to calculate the block-start border edge of the child.

Added a few new tests. Only two of them failed in NG prior to this CL,
but coverage seemed low, so I added them anyway. Two of them fail in the
legacy engine.

Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_layout_ng
Change-Id: I73b4aa02e5e88d1063f345db271cde1ad31ea9f5